+++ docs/HISTORY.rst
@@ -4,7 +4,8 @@ Changelog
 1.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Fixed links for title [kaeru]
+- Added keyword field to listing, and form [kaeru]
 
 
 1.0 (2014-01-29)