+++ docs/HISTORY.rst
@@ -4,8 +4,9 @@ Changelog
 1.1 (unreleased)
 ----------------
 
-- Fixed links for title [kaeru]
-- Added keyword field to listing, and form [kaeru]
+- Fixed links for title [kaerumy]
+- Added keyword field to listing, and form [kaerumy]
+- Add download csv feature to goodpracticelisting [kaerumy]
 
 
 1.0 (2014-01-29)