+++ docs/HISTORY.rst
@@ -6,6 +6,11 @@ Changelog
 
 - Nothing changed yet.
 
+1.2.2 (2015-10-08)
+------------------
+
+- add requirement for collective.excelexport
+
 
 1.2 (2014-02-10)
 ----------------